% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set.}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{tributary where fish was found}
  \item{MILE}{mile specifier for location where fish was found}
  \item{SPECIES}{species of fish}
  \item{LENGTH}{length of fish}
  \item{WEIGHT}{weight of fish}
  \item{DDT}{amount of DDT found in fish}
  ...
}
}
\source{
MATH4753 course textbook
}
\usage{
ddt
}
\description{
A dataset containing the DDT content and other attributes of 144
fish found in the Tennessee River system.
}
\keyword{datasets}
