% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from the fire}
  \item{DAMAGE}{damage of the fire}
  ...
}
}
\source{
MATH4753 course textbook
}
\usage{
fire
}
\description{
A dataset containing samples of fires and records the damage at a distance
from the fire
}
\keyword{datasets}
