% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2: Bootstraps a given sample}
\usage{
myboot2(iter, x, fun, alpha, cx, ...)
}
\arguments{
\item{iter}{how many times to run the sample of the given sample}

\item{x}{sample to bootstrap}

\item{fun}{function for the statistic to take from sample}

\item{alpha}{defines the quantile that is discarded from confidence interval}

\item{cx}{scales text and symbols in graph}

\item{...}{additional parameters for hist() function}
}
\description{
Creates a bootstrap interval of given statistic and plots a histogram of the vector of that statistic.
}
\details{
This function makes iter samples from given sample x and finds the
statistic (function for this statistic supplied as parameter) for each of the
samples created from x. A list of the confidence interval, the function, the
given sample, and the created vector of statistics is returned.
}
\examples{
\dontrun{
sam=rnorm(n=20,mean=10,sd=4)
myboot2(iter=10000,x=sam,fun="mean",alpha=0.20)
}

}
