% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu: Sample and graph from a uniform distribution}
\usage{
mycltu(n,iter,a,b)
}
\arguments{
\item{n}{sample size of each experiment}

\item{iter}{number of iterations of experiment}

\item{a}{lower bound of uniform distribution}

\item{b}{upper bound of uniform distribution}
}
\description{
Creates a random sample from a specified uniform distribution
and makes a histogram from sample means.
}
\details{
This function creates a random sample from a uniform distribution
from parameters a and b, each sample being size n, and there are iter iterations
of this experiment. The means of each iteration's samples will be returned
as a list and a histogram of the means is made.
}
\examples{
\dontrun{
mycltu(n=20,iter=100000)
}

}
