% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt: Plot and Dataframe by SPECIES}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{data frame, should hold the contents of DDT.csv}

\item{SPECIES}{only entries of this species is used in the plot}
}
\description{
Creates a plot, csv file, and a list terminal output for ddt
data filtered by species.
}
\details{
This function is made specifically for the DDT.csv data set,
and will print a LENGTH vs WEIGHT plot of the data points of the specified
SPECIES, colored by RIVER and with a quadratic curve of best fit. This
function also creates a .csv file named "LvsWforSPECIES.csv" with SPECIES
substituted with the actual SPECIES argument used for the function.
}
\examples{
\dontrun{
ddt=read.csv("DDT.CSV")
myddt(df=ddt, SPECIES="CCATFISH")
}

}
