% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve()}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{value for lower tail}
}
\description{
Plots lower tail curve area of given normal distribution and given value
}
\examples{
\dontrun{myncurve(30,10,0.7)}
}
