% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myrcurve()}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{iter}{number of iterations, default value 100}

\item{n}{sample size, default value 10}

\item{p}{probability, default value 0.5}
}
\value{
A barplot which tabulates each experiment's results
}
\description{
Creates and plots random sample trials of binary distributions.
}
\examples{
mybin(30,10,0.7)
}
